requirements.txt
Flask==3.0.0
Flask-Session==0.6.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
Werkzeug==3.0.1
python-dotenv==1.0.0
